+++ All.ui/miniprogram/tools/sideBox/sideBox.js
@@ -23,7 +23,7 @@ Component({
 
   ready: function() {
     this.setData({
-      height: app.globalData.statusBarHeight
+      height: app.globalData.bar
     })
   },